+++ src/engine/weapons/base_weapon.h
@@ -21,14 +21,14 @@ protected:
   // cppcheck-suppress unusedStructMember
   const uint8_t damage;
   // cppcheck-suppress unusedStructMember
-  const float cooldown;
+  const double cooldown;
   // cppcheck-suppress unusedStructMember
   const uint8_t ammo_speed;
   // cppcheck-suppress unusedStructMember
   int player_pos;
 
 public:
-  BaseWeapon(Snapshot &snap, uint16_t ammo, uint8_t damage, float cooldown,
+  BaseWeapon(Snapshot &snap, uint16_t ammo, uint8_t damage, double cooldown,
              uint8_t ammo_speed, int position);
   virtual Bullet shoot(Rectangle rectangle, uint8_t facing_direction,
                        ServerMap map) = 0;
